requirements.txt
aiohttp==3.6.2
aiohttp-jinja2==1.2.0
uvloop==0.14.0
aioreloader==0.3.0
pyYaml==5.3.1
asyncpg==0.20.1
asyncpgsa==0.26.1